+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Better handling of unsupported media types, allowing users to open them in their browsers
 - StatusMediaAttachment component for reusing the same logic across status media previews and the full viewer
 - transitionedComposable function which automatically applies correct transitions to the page
+- Notification filtering selection in notifications page
 
 ### Fixed
 - AMOLED theme has better contrast on some surfaces